backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
schwab-py==1.5.1
python-dotenv==1.0.0
websockets==12.0
aiohttp==3.9.1
